big refactor continues
