+++ barustenberg/src/plonk/proof_system/widgets/transition_widgets/mod.rs
@@ -1,2 +1,4 @@
 pub(crate) mod arithmetic_widget;
+pub(crate) mod containers;
+pub(crate) mod getters;
 pub(crate) mod transition_widget;
